Further speed up sam_parse_B_vals.
Previously it parses the B string twice.  Once to count commas and
allocate memory and once to fill out the memory.

Now the code reallocates periodically and thus only needs a single
pass.  The effect on large B arrays is significant.

                    2-pass     1-pass     develop
       gcc7    -O2: 5565824443 3299046885
       gcc7    -O3: 5779736469 3400756477
       gcc13   -O2: 5565893109 3086808341
       gcc13   -O3: 5392426978 3346007015 9724589000
       clang10 -O2: 5344657729 3465765165
       clang10 -O3: 5348030140 3460058513
       clang16 -O2: 4563321159 3374951558
       clang16 -O3: 4575986193 3311061338 6398268577

Speed instability was still observed by modifying code elsewhere so
this has been improved by splitting up the function and adding
function alignment requests.  We could achieve a similar result by
compiler options such as -falign-loops=32, but this affects all code
and we have not evaluated the impact elsewhere.