Fix cram_index_query to iterate correct for ref HTS_IDX_NOCOOR.
cram_index_query_last does a loop on cram_index_query with the
previous index entry in "from".  This scans to find the last
container.  If we're doing a query of ref "*" however it comes in as
reference HTS_IDX_NOCOOR (-2) and fails the refid matching check.

This makes cram_index_query_last now work again for region "*".