Ensure hFILE_scheme_handler open and isremote are set
Check that hfile plugins have supplied open() and
isremote() functions in their hFILE_scheme_handler struct,
and refuse to add them if not.  Failing to check this
could lead to an attempt to call a NULL pointer when
the interfaces are used.

Fix up the "crypt4gh-needed" scheme handler, which did not
supply isremote(); and "mem" which failed to supply open().

Thanks to John Marshall for suggested validation code
in hfile_add_scheme_handler().

Credit to OSS-Fuzz
Fixes oss-fuzz 67349