Check interval start to avoid overflowing bin numbers
Check start positions of query intervals against the maximum position
representable in the index's geometry, to avoid negative bin numbers
and the resulting infinite loops in the do...while loop.

Introduce hts_bin_maxpos() and hts_idx_maxpos(), and use them wherever the
maxpos calculation appears. (Leave the latter private, at least for now.)

Also change the existing end checks to <= as end is exclusive -- note it
is used as end-1 in the code guarded by the checks.