Add check for cpuid (for htscodecs)
Updates the htscodecs submodule to commit a7ce404:
 * Add configure checks for cpuid symbols

Adds similar configure checks to HTSlib, along with adjustments
to the Makefile and hts_probe_cc.sh for people who like to run
"make" without "./configure".  For both, if the cpuid check
fails, the SSE and AVX tests will be skipped as the features
they test for would not be used even if the corresponding code
is built.