Check that __builtin_cpu_supports("ssse3") works
This builtin doesn't understand ssse3 on some old versions of
clang.  To allow builds using these compilers, add a configure
check to ensure it works.

For make-only builds, we currently assume it does for X86
platforms.  As the problem only affected clang 3.7 and 3.8, both
of which are very old now, it seems reasonable to require use
of configure to support them.