Complete the refactoring of nibble2base_ssse3() as an ordinary source file
Build simd.c as a source file, so there is one copy of nibble2base_ssse3()
rather than potentially one per translation unit that uses it.

When everything was static, the names needed no prefix. However now that
there is a non-static function pointer, it needs an htslib-specific
prefix to avoid polluting the namespace. (In libhts.so it will usually
be a non-visible symbol, but in libhts.a it needs a prefix.)