Move __attribute__((target)) check to configure too
Configure checks for __builtin_cpu_supports("ssse3"), so the check
for __attribute__((target)) might as well be there too. This enables
this internal-only infrastructure to be removed from htslib/hts_defs.h,
which is primarily for defines needed in the public headers.

Add a configure check for __attribute__((constructor)) too, as we will
likely be using it on other SIMD-using platforms as well.