Fix two occurrences of undefined behaviour in sam.c
In bam_init1() we would initialize a new bam1_t using calloc().
Thereby all bytes are automatically set to zero including the pointer
members such as data. However, the C standard allows for NULL to
have a different bit pattern than all zeros. Hence, we need to zero
the fields explicitly.

In bam_aux_first() we first have to check whether b->data is a NULL
pointer before doing any arithmetic on it as that is undefined
behaviour according to the C standard. (Even NULL+0 is disallowed.)