Add ARM Neon nibble2base_neon() implementation
As ARM CPU capability instructions are privileged, we also need to
add cpu_supports_neon() and implement it to query for Neon/AdvSIMD
in various platform-dependent ways. (On 32-bit ARM, glibc helpfully
renamed the HWCAP_* macros to HWCAP_ARM_*, so accept both on Linux.)

32-bit ARM GCC erroneously does not define vst1q_u8_x2() (bug 71233,
fixed in v14.1 in this case), so avoid it on 32-bit ARM by writing
interleaved via vst2q_u8() instead.