Read whole lines at once in fai_retrieve()
Because fai_retrieve() is given only well-formatted input containing
lines of the same length, it already knows exactly where the base and
non-graphic characters are. So in general the interval to be read will
look like

    ......ATGCAT    (read last six bases and line terminator)
    ATGCATGCATGC    (read complete line including line terminator)
    ATGCATGCATGC    (read complete line including line terminator)
    ATGC........    (read first four base characters)

and can be read a line at a time instead of a character at a time,
with special handling for the partial first and last lines, and
discarding the terminator characters at the end of each line read.