Protect against negative POS fields in CRAM encoder.
We check pos against the ref end, and if that fits we then fetch
`&ref[apos]` to get the reference, but this understeps the array.

The fastest way to handle all this and other related errors is simply
to sanity check the input for out of bounds POS and bail out early.

Credit to OSS-Fuzz
Fixes oss-fuzz 70014