Improve indentation and program flow in DEMO.md
The "..." to indicate clipped out code is useful as it removes a lot
of the unnecessary boiler plate, especially error checking, and allows
us to drill down to the essense of what we're trying to teach.
However without indentation it's not clear what bits are being
removed.

Also there are times when the code removed is an error handler plus
a subsequent else clause, which gives the false impression with the
indentation that the if-clause has been negated.  Eg:

    ...
    //select required field alone, this is useful for CRAM alone
    if (hts_set_opt(infile, CRAM_OPT_REQUIRED_FIELDS, SAM_FLAG) < 0) {
    ...
       //read header
       in_samhdr = sam_hdr_read(infile);
    ...

Clearly the `sam_hdr_read` is not within the if-clause of
`hts_set_opt` returning < 0, but it's how it reads because "..."
hid too much.  (In the above case just indenting the ... and
un-indenting the hdr read call reverses the mental image.)

My approach here is to replace all `...` error handlers with `... // error`
and to indent them accordingly.

Also where we only have a one line thing, removing open curly braces
if we've cut the close curly brace gives a more sensible view.

None of this is really changing the code we're showing, but presenting
it in a more obvious manner to those that don't understand what is and
isn't an error case.