Remove undefined behavior in expression language.
The bit-wise AND, OR and XOR logic works on the overloaded double
(res->d), but casting this to int64_t for NAN is undefined behaviour.

Infact the current cast doesn't even yield the correct bit-wise layout
for NAN, as (long)d and *(long *)&d give different results for d==NAN.
However the result wasn't used anyway as it's promptly cleared again
in the subsequent `if (undef) hts_expr_val_undef(res)` code.  Hence
undefined or otherwise, it made no difference.