Fix gawk-ism in Makefile
The awk script used to extract htscodec's version string (if
git describe fails or returns a hash) used a three-argument
form of "match", which is a GNU extension.  Rewrite for
better compatibility with other awk implementations,
notably "mawk", which supplies awk in some Cirrus-CI images.

Also fix the cirrus-ci badge, which linked to the wrong location.