Fix possible 1 byte underflow in find_file_extension()
When checking for another delimiter before .gz or .bgz, the
pointer into the string was decremented without checking to
see if was at the start.  This could lead to a one byte read
before the start of the string when copying out the delimiter.

Credit to OSS_Fuzz
Fixes oss-fuzz id 71740