Ensure configure-less builds still enable NEON where available
Add lines to the Makefile-generated config.h so HAVE_GETAUXVAL
or HAVE_ELF_AUX_INFO is set when appropriate.

Add lines to Makefile and simd.c to set HAVE_ELF_AUX_INFO
on a recent enough OpenBSD.  Part of this check needs to be
in simd.c because it pulls in sys/param.h, which defines
some macros (notable MIN and MAX) that don't get on with the
rest of HTSlib, and so need to be limited in scope.  This
can be simplified later once OpenBSD versions without
elf_aux_info() are no longer supported.