Prevent segfault on empty tbi index
When an empty VCF file with a header but no data lines is indexed,
tbx_seqnames() called via bcf_sr_add_reader() returns an empty list.
Consequently, when bcf_sr_regions_next() is called and no in-memory
regions are available, it attempts to read from a file which is
not initialized.

This fixes https://github.com/samtools/bcftools/issues/2286