Fix minor inconsistency in test data.
test/sam's test_bam_set1_write_and_read_back.tmp.bam file had mate
information (rnext, pnext, tlen) for a record that was not paired.
This triggers a CRAM "fix" where some, but strangley not all, of this
was changed.

Arguably CRAM should correct all 3 or none, but regardless the test
data we are generating should be correct.