Add an hts_crc32 function to use zlib or libdeflate.
This follows on from the hts_md5* functions which wrap up either
OpenSSL or our own implementation.  Libdeflate's crc32 function is
considerably faster than the native zlib, so we want to use it in (for
example) the new "samtools checksum" code, but we do not wish to add
baggage of looking for libdeflate in the configure script.