Convert U to T instead of U to N when sam_parsing.
As this changes seq_nt16_table it changes it for all other uses of
this lookup table too, which will be widespread.  However this feels
like a reasonable thing to do given it only has an impact on data
which is currently out of bounds of what is expected.

Fixes samtools/samtools#2131