Fix cram_encode fuzzer issue caused by negative reference lengths.
Introduced in 24e4e31 with the cache of the LN_length field.  We now
validate this as positive.  The check in cram_encode.c should now not
be necessary, but it's fixed in two places incase we ever get this
cropping up via another route.

Credit to OSS_Fuzz
Fixes oss-fuzz issue 382922241