Add initial support for VCFv4.4 / VCFv4.5 Number= fields
Adds BCF_VL_* types for Number= P, LA, LG, LR and M

Upgrades bcf_hdr_register_hrec() so that it sets the appropriate
value when it finds one of the new types in a FORMAT header line.

Only set number type BCF_VL_FIXED if the value can actually be
parsed as an integer.  Unknown types will now cause a warning
to be printed and will be treated as BCF_VL_VAR.

Adds doxygen style comments to describe the meaning of the various
BCF_VL_* values.