Add hts_tpool_worker_id() API
This returns a numeric value from 0 to nthreads-1 corresponding to the
current running thread, or -1 if unthreaded or the thread does not
correspond to one allocated to this pool.

It may be used to associate data to a thread rather than to a job.
For example maintaining one open file handle per thread spawned.