Remove cram seek ability to do range queries via SEEK_CUR.
This was a feature that came all the way from the initial index
support added to io_lib, but I think it's a misfeature.  The
consequence of it is on failing with a SEEK_SET (eg network error,
or file corruption) it falls back to doing a read-and-discard loop to
simulate the seek via SEEK_CUR.

This may perhaps be of use when querying stdin, but it's highly
unlikely for us to be doing that while also having an index on disk
and it's not something we support with other formats.

Fixes #1877