Make HAVE_ATTRIBUTE_TARGET check also check SSSE3 intrinsics work
GCC 4.8.5 and earlier accept SSSE3 intrinsics with -mssse3; these
compiler versions also accept __attribute__((target("ssse3"))) but
the attribute fails to enable compilation of the intrinsics.

Refactor the check to HAVE_ATTRIBUTE_TARGET_SSSE3, and check both
the attribute and that the SSSE3 intrinsics used (in simd.c) are
in fact compilable in a function with the attribute. Fixes #1838.