Mention the CRAM 3.1 change.
I added this as an introductory paragraph rather than a bulleted item
given the significance.

I also link to out web pages, but this needs updating further with
modern data sets and to clarify the tooling so it's more than a set of
benchmarks.  This is something I'll work on separately.