Add ref-cache, a caching proxy for reference sequences
ref-cache is a local server of reference sequences, for use
when encoding or decoding CRAM files that use reference-based
compression.  It can also act as a proxy for an upstream server
(like the EBI's existing CRAM reference server) so that it
can download and cache new references as they are needed.

ref-cache is only built for MacOS, Linux, and other UNIX-like
platforms.  Windows is not supported due to the significant
differences in the platform.  It is also only built if
the configure script has been run.

As it's quite complex, and needs to deal with potentially
untrusted inputs, ref-cache is built with somewhat stricter
compiler options than the rest of HTSlib to try to reduce the
number of lurking bugs.  configure.ac is updated to add
tests for these options, along with the code to enable ref-cache.
These options are currently only enabled for compilers that
identify as gcc as it is unlikely that they are supported on
others.

ref-cache uses some APIs - notably openat(), mkdirat() etc. that
first appeared in POSIX 200809.  This requires _XOPEN_SOURCE
to be bumped from 600 to 700 to allow them to be used.

When built, ref-cache is tested by starting two servers, one
pre-populated with some files, and a second with no starting
files but with the first as its upstream.  If then attempts
to pull references from one to another by viewing an
example file with `test/test_view`, using the various servers as
the reference source.  This showed up some issues on the CI tests,
as they now actively use the hfile_libcurl plugin, which was
previously built but not really tested.  These are fixed by
ensuring HTS_PATH is set before running `test/test.pl`,
and by disabling the Address Sanitizer one-definition-rule check
in `test/with-shlib.sh`, as it otherwise triggers when the
static `test/test_view` program attempts to link in the
shared library `libhts.so`.