Fix possible buffer overruns in expand_path()
Not updating `len` in all cases, combined with strncpy()
padding out its destination with NULs, could result in an
overflow of the output `path` buffer.  This could also happen
if it was asked to expand a path with a negative or large enough
value in a `%s` directive (e.g. %-2s).

Fixed by adding extra validation, and using memcpy() which only
touches the bytes we want it to.