Add more nuance for header Number=? checking.
Previously we were gracious in accepting Number=. as it turns in in
several places in the wild.  However for VCF 4.4 and 4.5 we're more
strict as these aren't common place yet and we can complain louder
about errors.

The small caveat here is using 4.4 and 4.5 defined tags in an earlier
VCF version.  This happens when for example people wish to use local
alleles.  They've been used for many years, but only finally landed in
4.5.  Hence we see VCF 4.3 using these tags and Number=. as the
correct type code didn't exist in that version.

Also added phase-set tests which were accidentally missed on the
previous commit.