Add htscodecs submodule dependency to htscodecs/htscodecs/version.h
Add an order-only dependency on the htscodecs/htscodecs/version.h
Makefile rule to match the one on the htscodecs/htscodecs/%.h
pattern rule.  This ensures that the check to see if the submodule
is present runs for rules that have htscodecs/htscodecs/version.h
as an explicit dependency, and that the
htscodecs/htscodecs/version.h rule itself fails if the submodule
is not present.