Make bcf_remove_allele_set() work with local alleles
Local alleles work similarly to global ones, except the set
of alleles present can vary per sample.  The mapping from local
to global numbering is stored in the LAA field.  Removing an
allele can change both the global numbering and also local
numbering if one of the entries referred to the delete allele.

If LAA data is present, it is updated to reflect the new global
numbering if that has changed due to an allele being removed.  At
the same time a map is constructed storing data about removed
or renumbered local entries, in a similar way to the global map[]
array; and per-sample ref+alt counts are recorded to use when
calculating LR, LA and LG Number= fields.

Both the map[] and laa_map[] arrays store -1 for removed alleles,
allowing them to be used instead of kbs_exists() when checking
which items to delete.  It's then possible to use the same code
to update global Number= A,R,G and local Number= LA,LR,LG
data by either using the global map[] array or the local laa_map[]
entries for the current sample.