Make CRAM codecs with encoding E_BYTE_ARRAY do bounds checking on out.
The huffman code already had bounds checking, but
cram_byte_array_stop_decode_char only checked for read-overflow on the
CRAM block and not write-overflow on the destination.  That pointer is
fine as we've preallocated it to be big enough *provided* the CRAM
is internally consistent.

NB encoding E_BYTE_ARRAY_BLOCK is fine as these use block_append which
dynamically growns the block.