Permit better fuzzing detection and fix block size growth.
Our previous commit fixed a buffer overrun, but it was surprising
this wasn't found by the fuzz checker.  One reason for this is
pre-allocating and over-sizing the arrays we allocate as a speed
optimiser.  We disable this when fuzzing.

While reviewing this, also identified we were always growing rather
than setting the size per new alignment, causing excessive memory
usage.