Improved error handling in cram_compress_block2. (PR #1931)
If we learn a method is good to use but that method subsequently fails
to encode, we don't just fail in cram_compress_block2 any more.
Instead we restart the trial / adaption technique again to learn what
works instead.

This fixes an issue in CRAM encoding when we attempted to SIMD rANS
encode a block of Ns, but after RLE is applied the resultant buffer
was too small for SIMD to work.  (Also fixed in htscodecs)

Brought comment for cram_compress_block2() up to date.
The original was written when it only chose between zlib
strategies.  It now supports many more than that.

(cherry picked from commit 09c89d8f12b8367a5b869b563f5ae6f0888fb0f3)