Make bcf_remove_allele_set() work for INFO tags with multiples of A
The tags CICN, CIEND, CILEN, CIPOS, MEINFO, and METRANS are
defined as Number=. in the VCF 4.4+ specification, but with the
additional rule that they're really either 2 x Number=A for the
CI* tags or 4 x Number=A for the ME* ones.

Handle this by spotting them, and setting the multiple we expect
to see so values are treated as groups of two or four.