For rlen calculation, only use SVLEN for DEL symbolic alleles
PR #1897 updated the VCF rlen calculation in line with VCF 4.4,
including taking account of INFO SVLEN for non-INS alleles.  This
was not correct though as it included non-symbolic alleles if they
(incorrectly) had a non-missing SVLEN; and other types of symbolic
allele that do not affect the length with respect to the reference.
In fact, only deletions can have an affect on rlen, so this changes
the calculation to only use SVLEN for <DEL> alleles.

A bug is also fixed on big-endian platforms where INFO and FORMAT
values were being accessed incorrectly.  These are always stored
in little-endian order, so need to be converted to native ordering
when read.  The tests are updated to include values greater than
128 to ensure that this works correctly.